fix signup and authorize

+++ app/Http/Controllers/UserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\UsersStoreRequest;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class UserController extends Controller
 {
@@ -27,9 +28,14 @@ class UserController extends Controller
      */
     public function index(Request $request)
     {
-        $users = $this->user->getList($request->all());
-        $role = User::$role;
-        return view('admin.users.index', compact(['users', 'role']));
+        if (Gate::allows('admin')) {
+            $users = $this->user->getList($request->all());
+            $role = User::$role;
+            return view('admin.users.index', compact(['users', 'role']));
+        } else{
+            return redirect(route('home'));
+        }
+
     }
     
     /**
@@ -39,8 +45,12 @@ class UserController extends Controller
      */
     public function create()
     {
-        $role = User::$role;
-        return view('admin.users.create', compact('role'));
+        if (Gate::allows('admin')) {
+            $role = User::$role;
+            return view('admin.users.create', compact('role'));
+        } else{
+            return redirect(route('home'));
+        }
     }
     
     /**
@@ -51,9 +61,12 @@ class UserController extends Controller
      */
     public function store(UsersStoreRequest $request)
     {
-        
-        $this->user->storeData($request);
-        return redirect($request->url_back ?? route('admin.users.index'));
+        if (Gate::allows('admin')) {
+            $this->user->storeData($request);
+            return redirect($request->url_back ?? route('admin.users.index'));
+        } else{
+            return redirect(route('home'));
+        }
     }
     
     /**
@@ -64,9 +77,13 @@ class UserController extends Controller
      */
     public function show($id)
     {
-        $user = $this->user->find($id);
-        $role = User::$role;
-        return view('admin.users.show', compact(['user', 'role']));
+        if (Gate::allows('admin')) {
+            $user = $this->user->find($id);
+            $role = User::$role;
+            return view('admin.users.show', compact(['user', 'role']));
+        } else{
+            return redirect(route('home'));
+        }
     }
     
     /**
@@ -77,8 +94,12 @@ class UserController extends Controller
      */
     public function edit($id)
     {
-        $user = $this->user->find($id);
-        return view('admin.users.edit', compact('user'));
+        if (Gate::allows('admin')) {
+            $user = $this->user->find($id);
+            return view('admin.users.edit', compact('user'));
+        } else{
+            return redirect(route('home'));
+        }
     }
     
     /**
@@ -102,14 +123,18 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-        $user = $this->user->find($id);
-        if ($user->role == User::ADMIN_ROLE) {
-            $user->delete();
-        } elseif ($user->status == 1) {
-            $user->update(['status' => 0]);
-        } else {
-            $user->update(['status' => 1]);
+        if (Gate::allows('admin')) {
+            $user = $this->user->find($id);
+            if ($user->role == User::ADMIN_ROLE) {
+                $user->delete();
+            } elseif ($user->status == 1) {
+                $user->update(['status' => 0]);
+            } else {
+                $user->update(['status' => 1]);
+            }
+            return back();
+        } else{
+            return redirect(route('home'));
         }
-        return back();
     }
 }
